requirements.txt
munch==4.0.0
numpy==1.24.1
Pillow==9.2.0
Pillow==11.2.1
PyYAML==6.0
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.3.2
timm==0.3.2
torch==2.1.2+cu118
torch==2.4.1+cu118
torchvision==0.16.2+cu118
torchvision==0.19.1+cu118
